requirements.txt
matplotlib==3.5.3
numpy==1.23.3
pandas==1.4.4
Pillow~=9.2.0
rasterio==1.2.10
scikit_learn==1.1.3
seaborn==0.12.1
streamlit==1.15.1
torch==1.10.2
torchinfo~=1.7.1
scikit-learn~=1.1.3
torchvision~=0.11.3